lombok mybatis 게시판(board) CRUD API 만들기

+++ Spring_Boot/EXAMPLE_SPRING/src/main/java/com/example/demo/mvc/controller/BoardController.java
@@ -49,8 +49,9 @@ public class BoardController {
 	 * @param board
 	 */
 	@GetMapping("/save")
-	public void save(Board board) {
-		boardService.save(board);
+	public int save(Board parameter) {
+		boardService.save(parameter);
+		return parameter.getBoardSeq();
 	}
 
 
@@ -59,8 +60,13 @@ public class BoardController {
 	 * @param boardSeq
 	 */
 	@GetMapping("/delete/{boardSeq}")
-	public void delete(@PathVariable int boardSeq) {
+	public boolean delete(@PathVariable int boardSeq) {
+		Board board = boardService.get(boardSeq);
+		if(board == null) {
+			return false;
+		}
 		boardService.delete(boardSeq);
+		return true;
 	}
 	
 }
